pyproject.toml
[project]
name = "script-editor-app"
version = "0.1.0"
description = "A personal, Flask-based scriptwriting studio"
authors = [
    {name = "zarouz"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "flask (<2.3)",
    "flask-mongoengine (>=1.0.0,<2.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
    "jouvence (>=0.4.2,<0.5.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
